Added Settings fragment and Location Provider

+++ app/src/main/java/com/amg/weathered/data/db/entity/CurrentWeatherEntry.kt
@@ -41,7 +41,7 @@ data class CurrentWeatherEntry(
     val uvIndex: Int,
 
     @ColumnInfo(name="visibility")
-    val visibility: Int,
+    val visibility: Double,
 
     @ColumnInfo(name="weather_code")
     @SerializedName("weather_code")
